requirements.txt
numpy<2.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
fastai==2.7.13
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2
Pillow==9.5.0
python-dotenv==1.0.0 